Remove logdet.h/logdet.cpp, and migrate implementation to ops.h/ops.cpp
