--- torch_xla/csrc/ops/logdet.h
@@ -1,22 +1,0 @@
-#pragma once
-
-#include "torch_xla/csrc/ir.h"
-
-namespace torch_xla {
-namespace ir {
-namespace ops {
-
-class LogDet : public Node {
- public:
-  LogDet(const Value& input);
-
-  std::string ToString() const override;
-
-  NodePtr Clone(OpList operands) const override;
-
-  XlaOpVector Lower(LoweringContext* loctx) const override;
-};
-
-}  // namespace ops
-}  // namespace ir
-}  // namespace torch_xla
