+++ torch_xla/csrc/ops/ops.h
@@ -194,6 +194,8 @@ NodePtr Bernoulli(const Value& input, const Value& probability);
 
 NodePtr Take(const Value& input, const Value& index);
 
+NodePtr LogDet(const Value& input);
+
 }  // namespace ops
 }  // namespace ir
 }  // namespace torch_xla
