+++ torch_xla/csrc/tensor_methods.cpp
@@ -55,7 +55,6 @@
 #include "torch_xla/csrc/ops/leaky_relu_backward.h"
 #include "torch_xla/csrc/ops/linear_interpolation.h"
 #include "torch_xla/csrc/ops/log_softmax.h"
-#include "torch_xla/csrc/ops/logdet.h"
 #include "torch_xla/csrc/ops/masked_fill.h"
 #include "torch_xla/csrc/ops/masked_select.h"
 #include "torch_xla/csrc/ops/max_in_dim.h"
@@ -1370,7 +1369,7 @@ void XLATensor::log1p_(XLATensor& input) {
 }
 
 XLATensor XLATensor::logdet(const XLATensor& input) {
-  return input.CreateFrom(ir::MakeNode<ir::ops::LogDet>(input.GetIrValue()));
+  return input.CreateFrom(ir::ops::LogDet(input.GetIrValue()));
 }
 
 XLATensor XLATensor::lt(const XLATensor& input, at::Scalar other) {
